Swagger

+++ src/main/java/com/LabSupermarket/LabShoppingList/Service/CategoriaServiceImpl.java
@@ -5,14 +5,12 @@ import com.LabSupermarket.LabShoppingList.Entity.Produto;
 import com.LabSupermarket.LabShoppingList.Service.Interfaces.CategoriaService;
 import com.LabSupermarket.LabShoppingList.model.CategoriaRepository;
 import com.LabSupermarket.LabShoppingList.model.ProdutoRepository;
-import jakarta.persistence.EntityExistsException;
-import jakarta.persistence.EntityNotFoundException;
-import jdk.jshell.Snippet;
-import org.springframework.beans.BeanUtils;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
+
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
@@ -72,8 +70,12 @@ public class CategoriaServiceImpl implements CategoriaService {
 
     @Override
     public void excluir(Long id) {
+        Optional<Categoria> categoria = this.categoriaRepository.findById(id);
+        if (categoria.isPresent()) {
 
             this.categoriaRepository.deleteById(id);
-
+        } else {
+            throw new EntityNotFoundException("Esta Categoria está em uso e não pode ser deletada!");
+        }
     }
 }
